@@ -5,14 +5,6 @@
 import React from 'react';
 import createWidget from './lib/createWidget';
 
-let internalWidget;
-
-/**
- * Do nothing.
- */
-
-const noop = () => {};
-
 /**
  * Create a new `SoundCloud` component.
  *
@@ -20,53 +12,27 @@ const noop = () => {};
  * HTML5 widget from SoundCloud. Programmatic control not included.
  */
 
-const SoundCloud = React.createClass({
-  propTypes: {
-
-    // url to play. It's kept in sync, changing it will
-    // cause the widget to refresh and play the new url.
-    url: React.PropTypes.string.isRequired,
-
-    // custom ID for widget iframe element
-    id: React.PropTypes.string,
-
-    // widget parameters for appearance and auto play.
-    opts: React.PropTypes.objectOf(React.PropTypes.bool),
-
-    // event subscriptions
-    onPlay: React.PropTypes.func,
-    onPause: React.PropTypes.func,
-    onEnd: React.PropTypes.func
-  },
-
-  getDefaultProps() {
-    return {
-      id: 'react-sc-widget',
-      opts: {},
-      onPlay: noop,
-      onPause: noop,
-      onEnd: noop
-    };
-  },
+class SoundCloud extends React.Component {
+  constructor(props) {
+    super(props);
+    this._internalWidget = null;
+  }
 
   shouldComponentUpdate(nextProps) {
     return nextProps.url !== this.props.url;
-  },
+  }
 
   componentDidMount() {
-    createWidget(this.props.id, (widget) => {
-      this._setupWidget(widget);
-      this._reloadWidget();
-    });
-  },
+    this._createWidget();
+  }
 
   componentDidUpdate() {
     this._reloadWidget();
-  },
+  }
 
   componentWillUnmount() {
     this._unbindEvents();
-  },
+  }
 
   render() {
     return (
@@ -78,7 +44,18 @@ const SoundCloud = React.createClass({
               src='https://w.soundcloud.com/player/?url='
       />
     );
-  },
+  }
+
+  /**
+   * Create a new html5 widget and load it with some data.
+   */
+
+  _createWidget() {
+    createWidget(this.props.id, (widget) => {
+      this._setupWidget(widget);
+      this._reloadWidget();
+    });
+  }
 
   /**
    * Integrate a newly created `widget` with the rest of the component.
@@ -87,17 +64,17 @@ const SoundCloud = React.createClass({
    */
 
   _setupWidget(widget) {
-    internalWidget = widget;
+    this._internalWidget = widget;
     this._bindEvents();
-  },
+  }
 
   /**
    * Reload the embedded iframe with a new widget.
    */
 
   _reloadWidget() {
-    internalWidget.load(this.props.url, this.props.opts);
-  },
+    this._internalWidget.load(this.props.url, this.props.opts);
+  }
 
   /**
    * Listen for events coming from `widget`, and pass them up the
@@ -105,21 +82,46 @@ const SoundCloud = React.createClass({
    */
 
   _bindEvents() {
-    internalWidget.bind(window.SC.Widget.Events.PLAY, this.props.onPlay);
-    internalWidget.bind(window.SC.Widget.Events.PAUSE, this.props.onPause);
-    internalWidget.bind(window.SC.Widget.Events.FINISH, this.props.onEnd);
-  },
+    this._internalWidget.bind(window.SC.Widget.Events.PLAY, this.props.onPlay);
+    this._internalWidget.bind(window.SC.Widget.Events.PAUSE, this.props.onPause);
+    this._internalWidget.bind(window.SC.Widget.Events.FINISH, this.props.onEnd);
+  }
 
   /**
    * Remove all event bindings.
    */
 
   _unbindEvents() {
-    internalWidget.unbind(window.SC.Widget.Events.PLAY);
-    internalWidget.unbind(window.SC.Widget.Events.PAUSE);
-    internalWidget.unbind(window.SC.Widget.Events.FINISH);
+    this._internalWidget.unbind(window.SC.Widget.Events.PLAY);
+    this._internalWidget.unbind(window.SC.Widget.Events.PAUSE);
+    this._internalWidget.unbind(window.SC.Widget.Events.FINISH);
   }
-});
+}
+
+SoundCloud.propTypes = {
+  // url to play. It's kept in sync, changing it will
+  // cause the widget to refresh and play the new url.
+  url: React.PropTypes.string.isRequired,
+
+  // custom ID for widget iframe element
+  id: React.PropTypes.string,
+
+  // widget parameters for appearance and auto play.
+  opts: React.PropTypes.objectOf(React.PropTypes.bool),
+
+  // event subscriptions
+  onPlay: React.PropTypes.func,
+  onPause: React.PropTypes.func,
+  onEnd: React.PropTypes.func
+};
+
+SoundCloud.defaultProps = {
+  id: 'react-sc-widget',
+  opts: {},
+  onPlay: () => {},
+  onPause: () => {},
+  onEnd: () => {}
+};
 
 /**
  * Expose `SoundCloud` component
